refactor(bayn): make qualification decisions explicit
