+++ services/bayn/src/qualification-audit-command.ts
@@ -1,861 +1,51 @@
 import { NodeRuntime, NodeServices } from '@effect/platform-node'
-import { ClickhouseClient } from '@effect/sql-clickhouse'
-import { PgClient } from '@effect/sql-pg'
-import {
-  Config,
-  Context,
-  Data,
-  Effect,
-  FileSystem,
-  Layer,
-  Logger,
-  Redacted,
-  Result,
-  Schema,
-  Scope,
-  Stdio,
-  Stream,
-} from 'effect'
+import { Effect, Layer, Logger, Schema, Stdio, Stream } from 'effect'
 import * as Reactivity from 'effect/unstable/reactivity/Reactivity'
-import { ChildProcess, ChildProcessSpawner } from 'effect/unstable/process'
 
-import { EvaluationEventSchema, decodeInputManifestArtifact } from './evidence-contracts'
-import { MarketData, MarketDataLive, type MarketDataSnapshot } from './market-data'
-import { ProtocolSchema } from './protocol'
-import { QualificationLockSchema, QualificationResultSchema } from './qualification'
+import { qualificationAuditConfig, qualificationAuditCommandError } from './qualification/audit-command/model'
 import {
-  auditQualification,
-  classifySignalTableAccess,
-  validateSignalReplicaTopology,
-  type AuditDatabaseSnapshot,
-  type RepositoryAudit,
-  type SignalAccessRecord,
-} from './audit/audit'
-import { makeQualificationDossier } from './audit/dossier'
-import {
-  NonNegativeIntegerSchema as NonNegativeInteger,
-  PositiveIntegerSchema as PositiveInteger,
-  Sha256Schema as Sha256,
-  TrimmedNonEmptyStringSchema as NonEmptyString,
-  strictParseOptions as StrictParseOptions,
-} from './schemas'
-import type { Protocol, InputManifest } from './types'
+  liveQualificationAuditAcquirers,
+  makeQualificationAuditReaders,
+  runQualificationAudit,
+} from './qualification/audit-command/program'
 
-const IsoInstant = Schema.String.check(Schema.isPattern(/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{6}Z$/))
-const ReplicaName = NonEmptyString
-const GateScalar = Schema.Union([Schema.Finite, Schema.Boolean, Schema.String])
-const AuditClickhouseUrls = Config.Array(Schema.URLFromString).check(
-  Schema.isMinLength(2),
-  Schema.isMaxLength(8),
-  Schema.isUnique(),
-)
+export { readAuditDatabase } from './qualification/audit-command/database'
+export { loadAuditSignal } from './qualification/audit-command/signal'
+export { readAuditSignalAccess } from './qualification/audit-command/signal-access'
+export {
+  QualificationAuditCommandError,
+  type AcquireAuditDatabaseClient,
+  type AcquireAuditRepositoryClient,
+  type AcquireAuditSignalClient,
+  type AcquireAuditSignalReplicaClient,
+  type AuditConfig,
+  type AuditDatabaseClient,
+  type AuditRepositoryClient,
+  type AuditSignalClient,
+  type AuditSignalReplicaClient,
+  type QualificationAuditAcquirers,
+  type QualificationAuditReaders,
+  type SignalReplicaAccess,
+} from './qualification/audit-command/model'
+export { makeQualificationAuditReaders, runQualificationAudit } from './qualification/audit-command/program'
 
-const RunRow = Schema.Struct({
-  run_id: Sha256,
-  protocol_hash: Sha256,
-  snapshot_id: Sha256,
-  evaluation_schema_version: NonEmptyString,
-  source_revision: NonEmptyString,
-  image_repository: NonEmptyString,
-  image_digest: NonEmptyString,
-  strategy_name: Schema.Literal('risk-balanced-trend'),
-  initial_capital_micros: Schema.String,
-  status: Schema.Literal('COMPLETE'),
-  artifact_count: PositiveInteger,
-  event_count: NonNegativeInteger,
-  gate_count: PositiveInteger,
-  schema_version: NonEmptyString,
-  behavior_hash: Sha256,
-  parameter_hash: Sha256,
-  parameters: ProtocolSchema,
-})
-const ArtifactRow = Schema.Struct({
-  artifact_name: NonEmptyString,
-  schema_version: NonEmptyString,
-  content_hash: Sha256,
-  payload: Schema.Json,
-})
-const EventRow = Schema.Struct({
-  ordinal: NonNegativeInteger,
-  event_id: Sha256,
-  event_kind: Schema.Literals(['decision', 'fill', 'fee', 'cash-yield']),
-  content_hash: Sha256,
-  payload: EvaluationEventSchema,
-})
-const GateRow = Schema.Struct({
-  ordinal: NonNegativeInteger,
-  gate_name: NonEmptyString,
-  passed: Schema.Boolean,
-  actual: GateScalar,
-  required: GateScalar,
-  content_hash: Sha256,
-})
-const StatusRow = Schema.Union([
-  Schema.Struct({
-    status: Schema.Literal('WRITING'),
-    detail: Schema.Struct({
-      artifactCount: PositiveInteger,
-      eventCount: NonNegativeInteger,
-      gateCount: PositiveInteger,
-    }),
-  }),
-  Schema.Struct({
-    status: Schema.Literal('COMPLETE'),
-    detail: Schema.Struct({
-      reconciliationExact: Schema.Literal(true),
-      verdict: Schema.Literals(['PASS', 'FAIL_CLOSED']),
-    }),
-  }),
-])
-const TrialRow = Schema.Struct({ run_id: Sha256 })
-const QualificationRow = Schema.Struct({
-  lock_created_at: IsoInstant,
-  result_committed_at: IsoInstant,
-  lock_id: Sha256,
-  analysis_hash: Sha256,
-  result_hash: Sha256,
-  verdict: Schema.Literals(['QUALIFIED', 'REJECTED']),
-  lock_payload: QualificationLockSchema,
-  result_payload: QualificationResultSchema,
-})
-const ReadOnlyRow = Schema.Struct({ read_only: Schema.Boolean })
-const AccessRow = Schema.Struct({
-  replica: ReplicaName,
-  query_id: NonEmptyString,
-  query_start_time: IsoInstant,
-  user: NonEmptyString,
-  tables: Schema.Array(NonEmptyString),
-})
-const ReplicaRow = Schema.Struct({ replica: ReplicaName })
-
-const decodeRunRow = Schema.decodeUnknownEffect(Schema.Tuple([RunRow]), StrictParseOptions)
-const decodeArtifactRows = Schema.decodeUnknownEffect(Schema.Array(ArtifactRow), StrictParseOptions)
-const decodeEventRows = Schema.decodeUnknownEffect(Schema.Array(EventRow), StrictParseOptions)
-const decodeGateRows = Schema.decodeUnknownEffect(Schema.Array(GateRow), StrictParseOptions)
-const decodeStatusRows = Schema.decodeUnknownEffect(Schema.Array(StatusRow), StrictParseOptions)
-const decodeTrialRows = Schema.decodeUnknownEffect(Schema.Array(TrialRow), StrictParseOptions)
-const decodeQualificationRow = Schema.decodeUnknownEffect(Schema.Tuple([QualificationRow]), StrictParseOptions)
-const decodeReadOnlyRow = Schema.decodeUnknownEffect(Schema.Tuple([ReadOnlyRow]), StrictParseOptions)
-const decodeAccessRows = Schema.decodeUnknownEffect(Schema.Array(AccessRow), StrictParseOptions)
-const decodeReplicaRow = Schema.decodeUnknownEffect(Schema.Tuple([ReplicaRow]), StrictParseOptions)
-const decodeReplicaRows = Schema.decodeUnknownEffect(Schema.Array(ReplicaRow), StrictParseOptions)
 const encodeJson = Schema.encodeUnknownEffect(Schema.fromJsonString(Schema.Json))
 
-const config = Config.all({
-  output: Config.schema(Schema.Literals(['audit', 'dossier']), 'BAYN_AUDIT_OUTPUT').pipe(Config.withDefault('audit')),
-  runId: Config.schema(Sha256, 'BAYN_AUDIT_RUN_ID'),
-  postgresUrl: Config.redacted('BAYN_AUDIT_POSTGRES_URL'),
-  postgresTls: Config.boolean('BAYN_AUDIT_POSTGRES_TLS').pipe(Config.withDefault(false)),
-  postgresCaPath: Config.string('BAYN_AUDIT_POSTGRES_CA_PATH').pipe(Config.withDefault('')),
-  signalUrl: Config.string('BAYN_AUDIT_SIGNAL_URL'),
-  signalUsername: Config.string('BAYN_AUDIT_SIGNAL_USERNAME'),
-  signalPublisherUsername: Config.string('BAYN_AUDIT_SIGNAL_PUBLISHER_USERNAME'),
-  signalPassword: Config.redacted('BAYN_AUDIT_SIGNAL_PASSWORD'),
-  auditClickhouseUrls: Config.schema(AuditClickhouseUrls, 'BAYN_AUDIT_CLICKHOUSE_URLS'),
-  auditClickhouseUsername: Config.string('BAYN_AUDIT_CLICKHOUSE_USERNAME'),
-  auditClickhousePassword: Config.redacted('BAYN_AUDIT_CLICKHOUSE_PASSWORD'),
-  repositoryPath: Config.string('BAYN_AUDIT_REPOSITORY_PATH').pipe(Config.withDefault('.')),
-  operationTimeoutMs: Config.schema(PositiveInteger, 'BAYN_AUDIT_OPERATION_TIMEOUT_MS').pipe(
-    Config.withDefault(60_000),
-  ),
-})
-
-export type AuditConfig = Config.Success<typeof config>
-
-export class QualificationAuditCommandError extends Data.TaggedError('QualificationAuditCommandError')<{
-  readonly operation: 'audit' | 'configuration' | 'repository' | 'signal-access'
-  readonly message: string
-  readonly cause?: unknown
-}> {}
-
-const commandError = (
-  operation: QualificationAuditCommandError['operation'],
-  message: string,
-  cause?: unknown,
-): QualificationAuditCommandError => new QualificationAuditCommandError({ operation, message, cause })
-
-const readPostgresCertificate = (
-  input: AuditConfig,
-): Effect.Effect<string | undefined, QualificationAuditCommandError, FileSystem.FileSystem> => {
-  if (!input.postgresTls) return Effect.void.pipe(Effect.as(undefined))
-  if (input.postgresCaPath.length === 0) {
-    return Effect.fail(commandError('configuration', 'BAYN_AUDIT_POSTGRES_CA_PATH is required with TLS'))
-  }
-  return FileSystem.FileSystem.pipe(
-    Effect.flatMap((fileSystem) => fileSystem.readFileString(input.postgresCaPath)),
-    Effect.mapError((cause) =>
-      commandError('configuration', 'failed to read the qualification-audit PostgreSQL CA certificate', cause),
-    ),
-  )
-}
-
-const readDatabaseWithClient = Effect.fnUntraced(function* (sql: PgClient.PgClient, runId: string) {
-  return yield* sql.withTransaction(
-    Effect.gen(function* () {
-      yield* sql`SET TRANSACTION ISOLATION LEVEL REPEATABLE READ READ ONLY`
-      const [readOnly] = yield* decodeReadOnlyRow(
-        yield* sql`SELECT current_setting('transaction_read_only') = 'on' AS read_only`,
-      )
-      const [run] = yield* decodeRunRow(
-        yield* sql`
-              SELECT
-                run.run_id,
-                run.protocol_hash,
-                run.snapshot_id,
-                run.evaluation_schema_version,
-                run.source_revision,
-                run.image_repository,
-                run.image_digest,
-                run.strategy_name,
-                run.initial_capital_micros::text AS initial_capital_micros,
-                run.status,
-                (SELECT count(*)::integer FROM evaluation_artifacts WHERE run_id = run.run_id) AS artifact_count,
-                (SELECT count(*)::integer FROM evaluation_events WHERE run_id = run.run_id) AS event_count,
-                (SELECT count(*)::integer FROM gate_outcomes WHERE run_id = run.run_id) AS gate_count,
-                protocol.schema_version,
-                protocol.behavior_hash,
-                protocol.parameter_hash,
-                protocol.parameters
-              FROM evaluation_runs AS run
-              JOIN protocol_locks AS protocol USING (protocol_hash)
-              WHERE run.run_id = ${runId}
-            `,
-      )
-      const artifacts = yield* decodeArtifactRows(
-        yield* sql`
-            SELECT artifact_name, schema_version, content_hash, payload
-            FROM evaluation_artifacts
-            WHERE run_id = ${runId}
-            ORDER BY artifact_name
-          `,
-      )
-      const events = yield* decodeEventRows(
-        yield* sql`
-            SELECT ordinal, event_id, event_kind, content_hash, payload
-            FROM evaluation_events
-            WHERE run_id = ${runId}
-            ORDER BY ordinal
-          `,
-      )
-      const gates = yield* decodeGateRows(
-        yield* sql`
-            SELECT ordinal, gate_name, passed, actual, required, content_hash
-            FROM gate_outcomes
-            WHERE run_id = ${runId}
-            ORDER BY ordinal
-          `,
-      )
-      const statuses = yield* decodeStatusRows(
-        yield* sql`
-            SELECT status, detail
-            FROM status_history
-            WHERE run_id = ${runId}
-            ORDER BY sequence
-          `,
-      )
-      const trials = yield* decodeTrialRows(
-        yield* sql`
-            WITH target_lock AS (
-              SELECT created_at
-              FROM qualification_locks
-              WHERE candidate_run_id = ${runId}
-            )
-            SELECT run_id FROM (
-              SELECT trial.run_id
-              FROM qualification_trials AS trial
-              CROSS JOIN target_lock
-              WHERE trial.observed_at < target_lock.created_at
-              UNION
-              SELECT result.run_id
-              FROM qualification_results AS result
-              CROSS JOIN target_lock
-              WHERE result.committed_at < target_lock.created_at
-            ) AS prior_trials
-            ORDER BY run_id
-          `,
-      )
-      const [qualification] = yield* decodeQualificationRow(
-        yield* sql`
-              SELECT
-                to_char(lock.created_at AT TIME ZONE 'UTC', 'YYYY-MM-DD"T"HH24:MI:SS.US"Z"') AS lock_created_at,
-                to_char(result.committed_at AT TIME ZONE 'UTC', 'YYYY-MM-DD"T"HH24:MI:SS.US"Z"') AS result_committed_at,
-                lock.lock_id,
-                result.analysis_hash,
-                result.result_hash,
-                result.verdict,
-                lock.payload AS lock_payload,
-                result.payload AS result_payload
-              FROM qualification_locks AS lock
-              JOIN qualification_results AS result USING (lock_id)
-              WHERE lock.candidate_run_id = ${runId}
-            `,
-      )
-      return {
-        transactionReadOnly: readOnly.read_only,
-        protocol: {
-          protocolHash: run.protocol_hash,
-          schemaVersion: run.schema_version,
-          strategyName: run.strategy_name,
-          behaviorHash: run.behavior_hash,
-          parameterHash: run.parameter_hash,
-          parameters: run.parameters,
-        },
-        run: {
-          runId: run.run_id,
-          protocolHash: run.protocol_hash,
-          snapshotId: run.snapshot_id,
-          evaluationSchemaVersion: run.evaluation_schema_version,
-          sourceRevision: run.source_revision,
-          imageRepository: run.image_repository,
-          imageDigest: run.image_digest,
-          strategyName: run.strategy_name,
-          initialCapitalMicros: run.initial_capital_micros,
-          status: run.status,
-          artifactCount: run.artifact_count,
-          eventCount: run.event_count,
-          gateCount: run.gate_count,
-        },
-        artifacts: artifacts.map((row) => ({
-          name: row.artifact_name,
-          schemaVersion: row.schema_version,
-          contentHash: row.content_hash,
-          payload: row.payload,
-        })),
-        events: events.map((row) => ({
-          ordinal: row.ordinal,
-          id: row.event_id,
-          kind: row.event_kind,
-          contentHash: row.content_hash,
-          payload: row.payload,
-        })),
-        gates: gates.map((row) => ({
-          ordinal: row.ordinal,
-          name: row.gate_name,
-          passed: row.passed,
-          actual: row.actual,
-          required: row.required,
-          contentHash: row.content_hash,
-        })),
-        statuses,
-        priorTrialRunIds: trials.map((row) => row.run_id),
-        qualification: {
-          lockCreatedAt: qualification.lock_created_at,
-          resultCommittedAt: qualification.result_committed_at,
-          storedLockId: qualification.lock_id,
-          storedAnalysisHash: qualification.analysis_hash,
-          storedResultHash: qualification.result_hash,
-          storedVerdict: qualification.verdict,
-          lock: qualification.lock_payload,
-          result: qualification.result_payload,
-        },
-      } satisfies AuditDatabaseSnapshot
-    }),
-  )
-})
-
-export interface AuditDatabaseClient {
-  readonly read: (runId: string) => Effect.Effect<AuditDatabaseSnapshot, QualificationAuditCommandError>
-}
-
-export type AcquireAuditDatabaseClient<R> = (
-  input: AuditConfig,
-) => Effect.Effect<AuditDatabaseClient, QualificationAuditCommandError, Scope.Scope | R>
-
-const acquireAuditDatabaseClient: AcquireAuditDatabaseClient<FileSystem.FileSystem | Reactivity.Reactivity> = (input) =>
-  readPostgresCertificate(input).pipe(
-    Effect.flatMap((ca) =>
-      PgClient.make({
-        url: input.postgresUrl,
-        ssl: ca === undefined ? undefined : { ca, rejectUnauthorized: true },
-        applicationName: 'bayn-qualification-audit',
-        connectTimeout: input.operationTimeoutMs,
-        idleTimeout: '30 seconds',
-        maxConnections: 1,
-        minConnections: 0,
-        transformJson: false,
-      }).pipe(
-        Effect.mapError((cause) => commandError('audit', 'PostgreSQL read-only qualification audit failed', cause)),
-      ),
-    ),
-    Effect.map(
-      (sql): AuditDatabaseClient => ({
-        read: (runId) =>
-          readDatabaseWithClient(sql, runId).pipe(
-            Effect.mapError((cause) => commandError('audit', 'PostgreSQL read-only qualification audit failed', cause)),
-          ),
-      }),
-    ),
-  )
-
-const readAuditDatabaseWith = <R>(
-  input: AuditConfig,
-  runId: string,
-  acquireClient: AcquireAuditDatabaseClient<R>,
-): Effect.Effect<AuditDatabaseSnapshot, QualificationAuditCommandError, R> =>
-  Effect.scoped(acquireClient(input).pipe(Effect.flatMap((client) => client.read(runId))))
-
-export function readAuditDatabase(
-  input: AuditConfig,
-  runId: string,
-): Effect.Effect<AuditDatabaseSnapshot, QualificationAuditCommandError, FileSystem.FileSystem | Reactivity.Reactivity>
-export function readAuditDatabase<R>(
-  input: AuditConfig,
-  runId: string,
-  acquireClient: AcquireAuditDatabaseClient<R>,
-): Effect.Effect<AuditDatabaseSnapshot, QualificationAuditCommandError, R>
-export function readAuditDatabase<R>(
-  input: AuditConfig,
-  runId: string,
-  acquireClient?: AcquireAuditDatabaseClient<R>,
-): Effect.Effect<
-  AuditDatabaseSnapshot,
-  QualificationAuditCommandError,
-  R | FileSystem.FileSystem | Reactivity.Reactivity
-> {
-  return acquireClient === undefined
-    ? readAuditDatabaseWith(input, runId, acquireAuditDatabaseClient)
-    : readAuditDatabaseWith(input, runId, acquireClient)
-}
-
-const loadSignalWithClient = (
-  input: AuditConfig,
-  manifest: InputManifest,
-  protocol: Protocol,
-  sql: ClickhouseClient.ClickhouseClient,
-): Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, Scope.Scope> => {
-  const marketDataConfig = {
-    operationTimeoutMs: input.operationTimeoutMs,
-    clickhouse: {
-      url: input.signalUrl,
-      username: input.signalUsername,
-      password: input.signalPassword,
-      snapshotId: manifest.finalizedSnapshot.snapshotId,
-      publicationAsOf: manifest.finalizedSnapshot.asOfSession,
-      calendarVersion: manifest.finalizedSnapshot.calendarVersion,
-      bounds: manifest.bounds,
-    },
-  }
-  return Layer.build(
-    MarketDataLive(marketDataConfig, protocol).pipe(
-      Layer.provide(Layer.succeed(ClickhouseClient.ClickhouseClient, sql)),
-    ),
-  ).pipe(
-    Effect.flatMap((context) => Context.get(context, MarketData).load),
-    Effect.mapError((cause) => commandError('signal-access', 'Signal snapshot audit read failed', cause)),
-  )
-}
-
-export interface AuditSignalClient {
-  readonly load: (
-    manifest: InputManifest,
-    protocol: Protocol,
-  ) => Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, Scope.Scope>
-}
-
-export type AcquireAuditSignalClient<R> = (
-  input: AuditConfig,
-) => Effect.Effect<AuditSignalClient, QualificationAuditCommandError, Scope.Scope | R>
-
-const acquireAuditSignalClient: AcquireAuditSignalClient<Reactivity.Reactivity> = (input) =>
-  ClickhouseClient.make({
-    url: input.signalUrl,
-    username: input.signalUsername,
-    password: Redacted.value(input.signalPassword),
-    database: 'signal',
-    application: 'bayn-qualification-audit',
-    request_timeout: input.operationTimeoutMs,
-  }).pipe(
-    Effect.mapError((cause) => commandError('signal-access', 'Signal snapshot audit read failed', cause)),
-    Effect.map(
-      (sql): AuditSignalClient => ({
-        load: (manifest, protocol) => loadSignalWithClient(input, manifest, protocol, sql),
-      }),
-    ),
-  )
-
-const loadAuditSignalWith = <R>(
-  input: AuditConfig,
-  manifest: InputManifest,
-  protocol: Protocol,
-  acquireClient: AcquireAuditSignalClient<R>,
-): Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, R> =>
-  Effect.scoped(acquireClient(input).pipe(Effect.flatMap((client) => client.load(manifest, protocol))))
-
-export function loadAuditSignal(
-  input: AuditConfig,
-  manifest: InputManifest,
-  protocol: Protocol,
-): Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, Reactivity.Reactivity>
-export function loadAuditSignal<R>(
-  input: AuditConfig,
-  manifest: InputManifest,
-  protocol: Protocol,
-  acquireClient: AcquireAuditSignalClient<R>,
-): Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, R>
-export function loadAuditSignal<R>(
-  input: AuditConfig,
-  manifest: InputManifest,
-  protocol: Protocol,
-  acquireClient?: AcquireAuditSignalClient<R>,
-): Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, R | Reactivity.Reactivity> {
-  return acquireClient === undefined
-    ? loadAuditSignalWith(input, manifest, protocol, acquireAuditSignalClient)
-    : loadAuditSignalWith(input, manifest, protocol, acquireClient)
-}
-
-export interface SignalReplicaAccess {
-  readonly replica: string
-  readonly topology: readonly string[]
-  readonly access: readonly SignalAccessRecord[]
-}
-
-const readSignalReplicaAccessWithClient = (
-  input: AuditConfig,
-  url: URL,
-  database: AuditDatabaseSnapshot,
-  finalizedAt: string,
-  signalTables: InputManifest['tables'],
-  sql: ClickhouseClient.ClickhouseClient,
-): Effect.Effect<SignalReplicaAccess, QualificationAuditCommandError> => {
-  const barsTable = `signal.${signalTables.bars}`
-  const sessionsTable = `signal.${signalTables.sessions}`
-  const manifestTable = `signal.${signalTables.manifests}`
-  return Effect.gen(function* () {
-    const [replicaRow] = yield* decodeReplicaRow(
-      yield* sql`
-          SELECT host_name AS replica
-          FROM system.clusters
-          WHERE cluster = 'default' AND is_local
-        `,
-    )
-    const replica = replicaRow.replica
-    const topology = (yield* decodeReplicaRows(
-      yield* sql`
-        SELECT host_name AS replica
-        FROM system.clusters
-        WHERE cluster = 'default'
-        ORDER BY shard_num, replica_num
-      `,
-    )).map((row) => row.replica)
-    const rows = yield* sql`
-      SELECT
-        ${sql.param('String', replica)} AS replica,
-        query_id,
-        formatDateTime(toTimeZone(query_start_time_microseconds, 'UTC'), '%Y-%m-%dT%H:%i:%S.%fZ') AS query_start_time,
-        user,
-        tables
-      FROM system.query_log
-      WHERE type = 'QueryStart'
-        AND query_start_time_microseconds >= parseDateTime64BestEffort(${sql.param('String', finalizedAt)}, 6)
-        AND query_start_time_microseconds <= parseDateTime64BestEffort(
-          ${sql.param('String', database.qualification.resultCommittedAt)},
-          6
-        )
-        AND position(query, ${sql.param('String', database.run.snapshotId)}) > 0
-        AND (
-          has(tables, ${sql.param('String', manifestTable)})
-          OR has(tables, ${sql.param('String', sessionsTable)})
-          OR has(tables, ${sql.param('String', barsTable)})
-        )
-      ORDER BY query_start_time_microseconds, query_id
-    `.pipe(sql.withQueryId(`bayn-audit-access-${database.run.runId.slice(-24)}`))
-    const access: SignalAccessRecord[] = []
-    for (const row of yield* decodeAccessRows(rows)) {
-      const classification = classifySignalTableAccess(row.tables, signalTables)
-      if (Result.isFailure(classification)) {
-        return yield* Effect.fail(
-          commandError(
-            'signal-access',
-            'ClickHouse query-log row has no Signal evidence table',
-            classification.failure,
-          ),
-        )
-      }
-      access.push({
-        replica: row.replica,
-        queryId: row.query_id,
-        queryStartTime: row.query_start_time,
-        user: row.user,
-        kind: classification.success,
-      })
-    }
-    return { replica, topology, access }
-  }).pipe(Effect.mapError((cause) => commandError('signal-access', 'ClickHouse replica audit read failed', cause)))
-}
-
-export interface AuditSignalReplicaClient {
-  readonly url: URL
-  readonly read: (
-    database: AuditDatabaseSnapshot,
-    finalizedAt: string,
-    signalTables: InputManifest['tables'],
-  ) => Effect.Effect<SignalReplicaAccess, QualificationAuditCommandError>
-}
-
-export type AcquireAuditSignalReplicaClient<R> = (
-  input: AuditConfig,
-  url: URL,
-) => Effect.Effect<AuditSignalReplicaClient, QualificationAuditCommandError, Scope.Scope | R>
-
-const acquireAuditSignalReplicaClient: AcquireAuditSignalReplicaClient<Reactivity.Reactivity> = (input, url) =>
-  ClickhouseClient.make({
-    url: url.href,
-    username: input.auditClickhouseUsername,
-    password: Redacted.value(input.auditClickhousePassword),
-    database: 'system',
-    application: 'bayn-qualification-audit',
-    request_timeout: input.operationTimeoutMs,
-  }).pipe(
-    Effect.mapError((cause) => commandError('signal-access', 'ClickHouse replica audit read failed', cause)),
-    Effect.map(
-      (sql): AuditSignalReplicaClient => ({
-        url,
-        read: (database, finalizedAt, signalTables) =>
-          readSignalReplicaAccessWithClient(input, url, database, finalizedAt, signalTables, sql),
-      }),
-    ),
-  )
-
-const readSignalAccessWithClients = (
-  input: AuditConfig,
-  database: AuditDatabaseSnapshot,
-  finalizedAt: string,
-  signalTables: InputManifest['tables'],
-  clients: readonly AuditSignalReplicaClient[],
-): Effect.Effect<
-  { readonly replicas: readonly string[]; readonly access: readonly SignalAccessRecord[] },
-  QualificationAuditCommandError
-> =>
-  Effect.all(
-    clients.map((client) => client.read(database, finalizedAt, signalTables)),
-    { concurrency: Math.min(clients.length, 4) },
-  ).pipe(
-    Effect.flatMap((sources) =>
-      Effect.fromResult(validateSignalReplicaTopology(sources)).pipe(
-        Effect.mapError((cause) =>
-          commandError('signal-access', 'ClickHouse replica audit topology is invalid', cause),
-        ),
-      ),
-    ),
-  )
-
-const readAuditSignalAccessWith = <R>(
-  input: AuditConfig,
-  database: AuditDatabaseSnapshot,
-  finalizedAt: string,
-  signalTables: InputManifest['tables'],
-  acquireClient: AcquireAuditSignalReplicaClient<R>,
-): Effect.Effect<
-  { readonly replicas: readonly string[]; readonly access: readonly SignalAccessRecord[] },
-  QualificationAuditCommandError,
-  R
-> =>
-  Effect.scoped(
-    Effect.forEach(input.auditClickhouseUrls, (url) => acquireClient(input, url), {
-      concurrency: Math.min(input.auditClickhouseUrls.length, 4),
-    }).pipe(
-      Effect.flatMap((clients) => readSignalAccessWithClients(input, database, finalizedAt, signalTables, clients)),
-    ),
-  )
-
-export function readAuditSignalAccess(
-  input: AuditConfig,
-  database: AuditDatabaseSnapshot,
-  finalizedAt: string,
-  signalTables: InputManifest['tables'],
-): Effect.Effect<
-  { readonly replicas: readonly string[]; readonly access: readonly SignalAccessRecord[] },
-  QualificationAuditCommandError,
-  Reactivity.Reactivity
->
-export function readAuditSignalAccess<R>(
-  input: AuditConfig,
-  database: AuditDatabaseSnapshot,
-  finalizedAt: string,
-  signalTables: InputManifest['tables'],
-  acquireClient: AcquireAuditSignalReplicaClient<R>,
-): Effect.Effect<
-  { readonly replicas: readonly string[]; readonly access: readonly SignalAccessRecord[] },
-  QualificationAuditCommandError,
-  R
->
-export function readAuditSignalAccess<R>(
-  input: AuditConfig,
-  database: AuditDatabaseSnapshot,
-  finalizedAt: string,
-  signalTables: InputManifest['tables'],
-  acquireClient?: AcquireAuditSignalReplicaClient<R>,
-): Effect.Effect<
-  { readonly replicas: readonly string[]; readonly access: readonly SignalAccessRecord[] },
-  QualificationAuditCommandError,
-  R | Reactivity.Reactivity
-> {
-  return acquireClient === undefined
-    ? readAuditSignalAccessWith(input, database, finalizedAt, signalTables, acquireAuditSignalReplicaClient)
-    : readAuditSignalAccessWith(input, database, finalizedAt, signalTables, acquireClient)
-}
-
-const repositoryAuditWithClient = (
-  processes: ChildProcessSpawner.ChildProcessSpawner['Service'],
-  repositoryPath: string,
-  sourceRevision: string,
-  lockCreatedAt: string,
-  resultIdentity: readonly string[],
-): Effect.Effect<RepositoryAudit, QualificationAuditCommandError> =>
-  Effect.gen(function* () {
-    const exists = yield* processes.exitCode(
-      ChildProcess.make('git', ['-C', repositoryPath, 'cat-file', '-e', `${sourceRevision}^{commit}`]),
-    )
-    const ancestor = yield* processes.exitCode(
-      ChildProcess.make('git', ['-C', repositoryPath, 'merge-base', '--is-ancestor', sourceRevision, 'origin/main']),
-    )
-    const pattern = resultIdentity.join('|')
-    const lines = yield* processes.lines(
-      ChildProcess.make('git', [
-        '-C',
-        repositoryPath,
-        'log',
-        'origin/main',
-        `--before=${lockCreatedAt}`,
-        '--format=%H',
-        `-G${pattern}`,
-      ]),
-    )
-    const references = new Set(lines.filter((line) => /^[0-9a-f]{40}$/.test(line)))
-    return {
-      sourceCommitExists: Number(exists) === 0,
-      sourceCommitAncestorOfMain: Number(ancestor) === 0,
-      preLockResultReferences: [...references].sort(),
-    }
-  }).pipe(Effect.mapError((cause) => commandError('repository', 'repository provenance audit failed', cause)))
-
-export interface AuditRepositoryClient {
-  readonly audit: (
-    sourceRevision: string,
-    lockCreatedAt: string,
-    resultIdentity: readonly string[],
-  ) => Effect.Effect<RepositoryAudit, QualificationAuditCommandError>
-}
-
-export type AcquireAuditRepositoryClient<R> = (
-  input: AuditConfig,
-) => Effect.Effect<AuditRepositoryClient, QualificationAuditCommandError, R>
-
-const acquireAuditRepositoryClient: AcquireAuditRepositoryClient<ChildProcessSpawner.ChildProcessSpawner> = (input) =>
-  ChildProcessSpawner.ChildProcessSpawner.pipe(
-    Effect.map(
-      (processes): AuditRepositoryClient => ({
-        audit: (sourceRevision, lockCreatedAt, resultIdentity) =>
-          repositoryAuditWithClient(processes, input.repositoryPath, sourceRevision, lockCreatedAt, resultIdentity),
-      }),
-    ),
-  )
-
-export interface QualificationAuditReaders<R = never> {
-  readonly readDatabase: (runId: string) => Effect.Effect<AuditDatabaseSnapshot, QualificationAuditCommandError, R>
-  readonly loadSignal: (
-    manifest: InputManifest,
-    protocol: Protocol,
-  ) => Effect.Effect<MarketDataSnapshot, QualificationAuditCommandError, R>
-  readonly readSignalAccess: (
-    database: AuditDatabaseSnapshot,
-    finalizedAt: string,
-    signalTables: InputManifest['tables'],
-  ) => Effect.Effect<
-    { readonly replicas: readonly string[]; readonly access: readonly SignalAccessRecord[] },
-    QualificationAuditCommandError,
-    R
-  >
-  readonly auditRepository: (
-    sourceRevision: string,
-    lockCreatedAt: string,
-    resultIdentity: readonly string[],
-  ) => Effect.Effect<RepositoryAudit, QualificationAuditCommandError, R>
-}
-
-export interface QualificationAuditAcquirers<R> {
-  readonly database: AcquireAuditDatabaseClient<R>
-  readonly signal: AcquireAuditSignalClient<R>
-  readonly signalReplica: AcquireAuditSignalReplicaClient<R>
-  readonly repository: AcquireAuditRepositoryClient<R>
-}
-
-const liveQualificationAuditAcquirers: QualificationAuditAcquirers<
-  FileSystem.FileSystem | Reactivity.Reactivity | ChildProcessSpawner.ChildProcessSpawner
-> = {
-  database: acquireAuditDatabaseClient,
-  signal: acquireAuditSignalClient,
-  signalReplica: acquireAuditSignalReplicaClient,
-  repository: acquireAuditRepositoryClient,
-}
-
-export const makeQualificationAuditReaders = <R>(
-  input: AuditConfig,
-  acquirers: QualificationAuditAcquirers<R>,
-): QualificationAuditReaders<R> => ({
-  readDatabase: (runId) => readAuditDatabase(input, runId, acquirers.database),
-  loadSignal: (manifest, protocol) => loadAuditSignal(input, manifest, protocol, acquirers.signal),
-  readSignalAccess: (database, finalizedAt, signalTables) =>
-    readAuditSignalAccess(input, database, finalizedAt, signalTables, acquirers.signalReplica),
-  auditRepository: (sourceRevision, lockCreatedAt, resultIdentity) =>
-    acquirers
-      .repository(input)
-      .pipe(Effect.flatMap((repository) => repository.audit(sourceRevision, lockCreatedAt, resultIdentity))),
-})
-
-export const runQualificationAudit = <R>(input: AuditConfig, readers: QualificationAuditReaders<R>) =>
-  Effect.gen(function* () {
-    const database = yield* readers.readDatabase(input.runId)
-    const inputManifestArtifact = database.artifacts.find((artifact) => artifact.name === 'input-manifest')
-    if (inputManifestArtifact === undefined) {
-      return yield* commandError('audit', 'input-manifest artifact is missing')
-    }
-    const manifest = yield* decodeInputManifestArtifact(inputManifestArtifact.payload)
-    const protocol = database.protocol.parameters
-    const signal = yield* readers.loadSignal(manifest, protocol)
-    const signalAccess = yield* readers.readSignalAccess(
-      database,
-      manifest.finalizedSnapshot.finalizedAt,
-      manifest.tables,
-    )
-    const result = database.qualification.result
-    const repository = yield* readers.auditRepository(
-      database.run.sourceRevision,
-      database.qualification.lockCreatedAt,
-      [database.run.runId, result.resultHash, result.analysis.analysisHash],
-    )
-    const auditInput = {
-      bars: signal.bars,
-      manifest: signal.manifest,
-      protocol,
-      database,
-      signalReplicas: signalAccess.replicas,
-      signalAccess: signalAccess.access,
-      signalPrincipals: { candidate: input.signalUsername, publishers: [input.signalPublisherUsername] },
-      repository,
-    }
-    return input.output === 'dossier'
-      ? yield* Effect.fromResult(makeQualificationDossier(auditInput)).pipe(
-          Effect.mapError((cause) => commandError('audit', 'qualification dossier construction failed', cause)),
-        )
-      : yield* Effect.fromResult(auditQualification(auditInput)).pipe(
-          Effect.mapError((cause) => commandError('audit', 'qualification audit evaluation failed', cause)),
-        )
-  })
-
 const main = Effect.gen(function* () {
-  const input = yield* config
+  const input = yield* qualificationAuditConfig
   const report = yield* runQualificationAudit(
     input,
     makeQualificationAuditReaders(input, liveQualificationAuditAcquirers),
   )
   const output = yield* encodeJson(report).pipe(
-    Effect.mapError((cause) => commandError('audit', 'qualification audit output encoding failed', cause)),
+    Effect.mapError((cause) =>
+      qualificationAuditCommandError('audit', 'qualification audit output encoding failed', cause),
+    ),
   )
   const stdio = yield* Stdio.Stdio
   yield* Stream.run(Stream.make(`${output}\n`), stdio.stdout())
   if (input.output === 'audit' && 'status' in report && report.status !== 'PASS') {
-    return yield* commandError('audit', 'qualification audit failed')
+    return yield* qualificationAuditCommandError('audit', 'qualification audit failed')
   }
 })
